Separate Location and Bluetooth permissions
